Fix issue with setting hint button to active...
which caused hints to be accessible even when formally depleted if the user
navigated away from the puzzle grid (e.g. to home page) and then back again.

Turns out the side-effect on LabelButton.active change was not needed, it's just
that the MouseArea needed to reference the parent's active property, presumably
because it already had active defined in its own namespace.